+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # lakeql-cli
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - lakeql-core@0.5.0
+  - lakeql-parquet@0.4.0
+  - lakeql-sql@0.2.1
+
 ## 0.0.8
 
 ### Patch Changes