+++ packages/fs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lakeql-fs
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - lakeql-core@0.5.0
+
 ## 0.0.5
 
 ### Patch Changes