+++ packages/s3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lakeql-s3
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - lakeql-core@0.5.0
+
 ## 0.0.8
 
 ### Patch Changes