+++ packages/sql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lakeql-sql
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - lakeql-core@0.5.0
+
 ## 0.2.0
 
 ### Minor Changes